Move UrlBuilder class into utility package
